tests/wpt/tests/clipboard-apis/permissions/writeText-denied.https.html

<!doctype html>
<meta charset="utf-8">
<title>navigator.clipboard.writeText() fails when permission denied</title>
<link rel="help" href="https://w3c.github.io/clipboard-apis/#async-clipboard-api">
<body>Body needed for test_driver.click()</body>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
<script src="../resources/user-activation.js"></script>
<script>
'use strict';

promise_test(async t => {
  await trySetPermission('clipboard-write', 'denied');
  await waitForUserActivation();
  await promise_rejects_dom(t, 'NotAllowedError',
      navigator.clipboard.writeText('xyz'));
}, 'navigator.clipboard.writeText() fails when permission denied');
</script>
